🔧 chore(potential): adapt to upstream's namespace-package refactor
Rebasing onto the current upstream main (post #799/#804/#805) surfaced
a few things the rebase itself couldn't fix, since these files are new
on this branch and had no conflicting upstream counterpart to merge
against:

- galax._custom_types -> galax.potential.custom_types and
  galax._interop.optional_deps -> galax.interop.optional_deps, matching
  the renamed modules.
- Added # type: ignore[no-any-return] to match the stricter mypy
  config that landed alongside the refactor, mirroring the existing
  NFW files.
- Dropped a now-unused # type: ignore[misc] on GaussLegendreIntegrator.
