+++ src/galax/potential/_src/builtin/gaussian/axisymmetric.py
@@ -177,13 +177,13 @@ class AxisymmetricGaussianPotential(AbstractSinglePotential):
         def integrand(s: Float[Array, "N"]) -> Float[Array, "N *batch"]:
             s2 = s.reshape(s.shape + (1,) * batchdims) ** 2
             denom = jnp.sqrt((qsq - 1) * s2 + 1)
-            return delta_psi_factor(s2) / denom
+            return delta_psi_factor(s2) / denom  # type: ignore[no-any-return]
 
         # TODO: option to do integrate.quad
         integral = self._integrator(integrand)
 
         out = (-2.0 * jnp.pi * self.constants["G"] * rho0 * r_s**2 * q2) * integral
-        return out.ustrip(self.units["specific energy"])
+        return out.ustrip(self.units["specific energy"])  # type: ignore[no-any-return]
 
     # ==========================================================================
 
@@ -203,4 +203,4 @@ class AxisymmetricGaussianPotential(AbstractSinglePotential):
         xi2 = self._spheroid_surface(xyz[None], qsq, s2)[0] / r_s**2
 
         dens = rho0 * jnp.exp(-xi2 / 2)
-        return dens.ustrip(self.units["mass density"])
+        return dens.ustrip(self.units["mass density"])  # type: ignore[no-any-return]
