+++ src/galax/potential/_src/builtin/gaussian/triaxial.py
@@ -211,13 +211,13 @@ class TriaxialGaussianPotential(AbstractSinglePotential):
         def integrand(s: Float[Array, "N"]) -> Float[Array, "N *batch"]:
             s2 = s.reshape(s.shape + (1,) * batchdims) ** 2
             denom = jnp.sqrt(((q1sq - 1) * s2 + 1) * ((q2sq - 1) * s2 + 1))
-            return delta_psi_factor(s2) / denom
+            return delta_psi_factor(s2) / denom  # type: ignore[no-any-return]
 
         # TODO: option to do integrate.quad
         integral = self._integrator(integrand)
 
         out = (-2.0 * jnp.pi * self.constants["G"] * rho0 * r_s**2 * q1 * q2) * integral
-        return out.ustrip(self.units["specific energy"])
+        return out.ustrip(self.units["specific energy"])  # type: ignore[no-any-return]
 
     # ==========================================================================
 
@@ -237,4 +237,4 @@ class TriaxialGaussianPotential(AbstractSinglePotential):
         xi2 = self._ellipsoid_surface(xyz[None], q1sq, q2sq, s2)[0] / r_s**2
 
         dens = rho0 * jnp.exp(-xi2 / 2)
-        return dens.ustrip(self.units["mass density"])
+        return dens.ustrip(self.units["mass density"])  # type: ignore[no-any-return]
