+++ tests/unit/potential/builtin/test_common.py
@@ -8,8 +8,8 @@ from plum import convert
 import quaxed.numpy as jnp
 import unxt as u
 
-import galax.potential.custom_types as gt
 import galax.potential as gp
+import galax.potential.custom_types as gt
 import galax.potential.params as gpp
 from ..param.test_field import ParameterFieldMixin
