+++ tests/unit/potential/builtin/test_triaxialgaussian.py
@@ -7,8 +7,8 @@ import pytest
 import quaxed.numpy as jnp
 import unxt as u
 
-import galax.potential.custom_types as gt
 import galax.potential as gp
+import galax.potential.custom_types as gt
 from ..test_core import AbstractSinglePotential_Test
 from .test_common import (
     ParameterMTotMixin,
@@ -17,7 +17,7 @@ from .test_common import (
     ParameterShapeQ2Mixin,
     assert_gaussian_matches_galpy,
 )
-from galax._interop.optional_deps import OptDeps
+from galax.interop.optional_deps import OptDeps
 
 
 class TestTriaxialGaussianPotential(
